+++ packages/server/lib/engine_io_server.dart
@@ -8,7 +8,6 @@ export 'package:engine_io_shared/transports.dart' show ConnectionType;
 
 export 'src/transports/polling/polling.dart' show PollingTransport;
 export 'src/transports/websocket/websocket.dart' show WebSocketTransport;
-export 'src/transports/heartbeat_manager.dart' show HeartbeatManager;
 export 'src/transports/transport.dart' show Transport;
 export 'src/configuration.dart' show ConnectionOptions, ServerConfiguration;
 export 'src/server.dart' show Server;