+++ packages/shared/CHANGELOG.md
@@ -15,6 +15,7 @@
   - Define the `Transport` class:
     - Define the `PollingTransport` mixin.
     - Define the `WebSocketTransport` mixin.
+  - Define the `Heart` class to manage the heartbeat mechanism.
   - Define the available connection types in the `ConnectionType` enum.
 - Options:
   - Define connection options in the `ConnectionOptions` class.